bdev: change blocklen from uint64_t to uint32_t
This is the size of a logical block in bytes; 4 GB is more than plenty.

Also allows cleaning up casts to uint32_t in the SCSI translation layer.

+++ include/spdk/bdev.h
@@ -107,7 +107,7 @@ struct spdk_bdev {
 	char product_name[SPDK_BDEV_MAX_PRODUCT_NAME_LENGTH];
 
 	/** Size in bytes of a logical block for the backend */
-	uint64_t blocklen;
+	uint32_t blocklen;
 
 	/** Number of blocks */
 	uint64_t blockcnt;
